+++ Changelog.md
@@ -4,6 +4,7 @@
 
 - Update Kamelet Catalog from 3.20.2 to 3.20.3
 - Provide command to create a Camel route defined in Java DSL using Camel JBang
+- Provide command to create a Camel route defined in XML DSL using Camel JBang
 
 ## 0.7.0