+++ Changelog.md
@@ -2,7 +2,7 @@
 
 ## 0.8.0
 
-- Update Kamelet Catalog from 3.20.2 to 3.20.3
+- Update Kamelet Catalog from 3.20.2 to 3.20.4
 - Provide command to create a Camel route defined in Java DSL using Camel JBang
 - Provide command to create a Camel route defined in XML DSL using Camel JBang